refactor author block for #504
