+++ src/blocks/author/index.js
@@ -4,146 +4,41 @@
  */
 import './styles/style.scss';
 import './styles/editor.scss';
-import Edit from './components/edit';
+import edit from './edit';
 import icons from './../../utils/icons';
+import transforms from './transforms';
+import save from './save';
+import metadata from './block.json';
 
 /**
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
-const { createBlock } = wp.blocks;
-const { RichText, InnerBlocks } = wp.blockEditor;
 
 /**
  * Block constants
  */
-const name = 'author';
+const { attributes, name } = metadata;
 
 const title = __( 'Author' );
 
 const icon = icons.author;
 
-const keywords = [
-	__( 'biography' ),
-	__( 'profile' ),
-	__( 'coblocks' ),
-];
-
-const blockAttributes = {
-	biography: {
-		type: 'array',
-		source: 'children',
-		selector: '.wp-block-coblocks-author__biography',
-		default: [],
-	},
-	heading: {
-		type: 'string',
-		selector: '.wp-block-coblocks-author__heading',
-		default: __( 'Written by...' ),
-	},
-	name: {
-		type: 'string',
-		selector: '.wp-block-coblocks-author__name',
-	},
-	imgId: {
-		type: 'number',
-	},
-	imgUrl: {
-		type: 'string',
-		source: 'attribute',
-		attribute: 'src',
-		selector: 'img',
-	},
-	textAlign: {
-		type: 'string',
-	},
-};
-
 const settings = {
 
-	title: title,
+	title,
 
 	description: __( 'Add an author biography.' ),
 
-	keywords: keywords,
+	keywords: [	__( 'biography' ),	__( 'profile' ), __( 'coblocks' ) ],
 
-	attributes: blockAttributes,
+	attributes,
 
-	transforms: {
-		from: [
-			{
-				type: 'raw',
-				selector: 'div.wp-block-coblocks-author',
-				schema: {
-					div: {
-						classes: [ 'wp-block-coblocks-author' ],
-					},
-				},
-			},
-			{
-				type: 'prefix',
-				prefix: ':author',
-				transform: function( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-					} );
-				},
-			},
-		],
-	},
+	transforms,
 
-	edit: Edit,
+	edit,
 
-	save( { attributes } ) {
-		const {
-			biography,
-			heading,
-			imgUrl,
-			name,
-			textAlign,
-		} = attributes;
-
-		if ( name ) {
-			return (
-				<div style={ { textAlign: textAlign } }>
-					{ imgUrl && (
-						<div className={ 'wp-block-coblocks-author__avatar' }>
-							<img
-								className="wp-block-coblocks-author__avatar-img"
-								src={ imgUrl }
-								alt="avatar"
-							/>
-						</div>
-					) }
-					<div className={ 'wp-block-coblocks-author__content' }>
-						{ ! RichText.isEmpty( heading ) && (
-							<RichText.Content
-								tagName="p"
-								className="wp-block-coblocks-author__heading"
-								value={ heading }
-							/>
-						) }
-						{ ! RichText.isEmpty( name ) && (
-							<RichText.Content
-								tagName="span"
-								className="wp-block-coblocks-author__name"
-								value={ name }
-							/>
-						) }
-						{ ! RichText.isEmpty( biography ) && (
-							<RichText.Content
-								tagName="p"
-								className="wp-block-coblocks-author__biography"
-								value={ biography }
-							/>
-						) }
-						<InnerBlocks.Content />
-					</div>
-				</div>
-			);
-		}
-		return null;
-	},
+	save,
 };
 
 export { name, title, icon, settings };
